Fix light ray epsilon for two-sided objects
Now that objects are two-sided surfaces, we need to move the light
ray starting point out from the surface slightly to avoid going
through it to the wrong side. This caused light to be cast on
the shadowed side of objects.

This is slightly ugly, but can't be helped if we want two-sided
objects...